all of ghc but not Prelude
I've copied all of /opt/ghc/7.10.3/lib/ghc-7.10.3 into the ghc-less
machine, expecting it would be everything needed to run hint and then
some, because that folder includes the ghc binary and a large number of
compiled library files. But it looks like hint also requires gcc to be
installed, not just ghc!