installing g++ on the target machine
Why g++ and not gcc? Because that's what the haskell image installs and
I want to play it safe.

Still doesn't work, I get an error message about libgmp.so not being
found. Maybe the ghc library needs to pass some flags to gcc to tell it
what libgmp.so is in c-libs?