use the same absolute path for the local database
It turns out the package-db configuration files hardcode a particular
absolute path, so we'll need to rewrite those in order to relocate the
database. For now, just to demonstrate that this can work, let's just
move the local database to the place it is expected, in the same way
we're moving the global database to the absolute path which hint
expects.